Fix for resetting kernels_table
kernels_table is a global table to populate all the available kernels.
This table is bound to the current setup api so it needs to be cleared
by its destroy api. Otherwise the old data will be preserved across
different setups and it will result in wrong execution of kernels.

Change-Id: If2285241109d8e35661f012fe95be1760de66881